refactor: add parse json query interceptor
