+++ src/common/application/errors/application-errors.ts
@@ -9,12 +9,11 @@ enum APPLICATION_ERROR {
   EMAIL_NOT_VERIFIED = 'EMAIL_NOT_VERIFIED',
   BAD_REQUEST = 'BAD_REQUEST',
   INVALID_SEARCH_OPERATOR = 'INVALID_SEARCH_OPERATOR',
-  INVALID_SEARCH_JSON = 'INVALID_SEARCH_JSON',
   INVALID_SEARCH_FIELD = 'INVALID_SEARCH_FIELD',
   INVALID_SEARCH_VALUE = 'INVALID_SEARCH_VALUE',
-  INVALID_ORDER_JSON = 'INVALID_ORDER_JSON',
   INVALID_ORDER_FIELD = 'INVALID_ORDER_FIELD',
   INVALID_ORDER_DIRECTION = 'INVALID_ORDER_DIRECTION',
+  INVALID_JSON_QUERY = 'INVALID_JSON_QUERY',
 }
 
 export class ApplicationError extends Error {
@@ -106,14 +105,6 @@ export class InvalidSearchOperatorError extends ApplicationError {
   }
 }
 
-export class InvalidSearchJsonError extends ApplicationError {
-  constructor() {
-    super('Invalid search json');
-    this.httpStatus = 400;
-    this.error = APPLICATION_ERROR.INVALID_SEARCH_JSON;
-  }
-}
-
 export class InvalidSearchFieldError extends ApplicationError {
   constructor() {
     super('Invalid search field');
@@ -130,14 +121,6 @@ export class InvalidSearchValueError extends ApplicationError {
   }
 }
 
-export class InvalidOrderJsonError extends ApplicationError {
-  constructor() {
-    super('Invalid order json');
-    this.httpStatus = 400;
-    this.error = APPLICATION_ERROR.INVALID_ORDER_JSON;
-  }
-}
-
 export class InvalidOrderFieldError extends ApplicationError {
   constructor() {
     super('Invalid order field');
@@ -151,5 +134,13 @@ export class InvalidOrderDirectionError extends ApplicationError {
     super('Invalid order direction');
     this.httpStatus = 400;
     this.error = APPLICATION_ERROR.INVALID_ORDER_DIRECTION;
+  }
+}
+
+export class InvalidJsonQueryError extends ApplicationError {
+  constructor() {
+    super('Invalid JSON query');
+    this.httpStatus = 400;
+    this.error = APPLICATION_ERROR.INVALID_JSON_QUERY;
   }
 }
